# dun_render: Unroll triangle loops